LC solved 1672 - Richest Customer Wealth
I solved this problem by iterating through the 2D array and calculating the sum of each row. I then compared the sum with the current max sum and updated the max sum if the current sum was greater. I returned the max sum as the result. The time complexity of this solution is O(n*m).